[CS-4087]: Migrate app to ts and functional comp
