+++ cardstack/src/types/react-native-dotenv.d.ts
@@ -22,4 +22,5 @@ declare module 'react-native-dotenv' {
   export const VERSION_TAP_COUNT: string;
   export const OPENSEA_API_KEY: string;
   export const SECURE_STORE_KEY: string;
+  export const SENTRY_ENDPOINT: string;
 }
